requirements.txt
numpy==1.23.5
python-dotenv==0.21.0
scipy==1.9.3
transformers==4.24.0
tweepy==4.12.1
